fix: Decompress Redis data when streaming.
